// src/payloads/features/idpay/onboarding/types.ts
export enum IDPayServiceID {
  DEFAULT = "TESTSRV01",
  INVITED = "TESTSRV02",
  NO_PREREQUISITES = "TESTSRV03",
  PDND_ONLY = "TESTSRV04",
  SELF_ONLY = "TESTSRV05",

  ERR_STATUS_NOT_ELIGIBLE = "TESTSRV06",
  ERR_STATUS_NO_REQUIREMENTS = "TESTSRV07",
  ERR_STATUS_ONBOARDED = "TESTSRV08",
  ERR_STATUS_UNSUBSCRIBED = "TESTSRV09",
  ERR_STATUS_ON_EVALUATION = "TESTSRV10",

  ERR_CHECK_BUDGET_TERMINATED = "TESTSRV11",
  ERR_CHECK_ENDED = "TESTSRV12",
  ERR_CHECK_NOT_STARTED = "TESTSRV13",
  ERR_CHECK_SUSPENDED = "TESTSRV14"
}

export enum IDPayInitiativeID {
  DEFAULT = "TESTINIT01",
  INVITED = "TESTINIT02",
  NO_PREREQUISITES = "TESTINIT03",
  PDND_ONLY = "TESTINIT04",
  SELF_ONLY = "TESTINIT05",

  ERR_STATUS_NOT_ELIGIBLE = "TESTINIT06",
  ERR_STATUS_NO_REQUIREMENTS = "TESTINIT07",
  ERR_STATUS_ONBOARDED = "TESTINIT08",
  ERR_STATUS_UNSUBSCRIBED = "TESTINIT09",
  ERR_STATUS_ON_EVALUATION = "TESTINIT10",

  ERR_CHECK_BUDGET_TERMINATED = "TESTINIT11",
  ERR_CHECK_ENDED = "TESTINIT12",
  ERR_CHECK_NOT_STARTED = "TESTINIT13",
  ERR_CHECK_SUSPENDED = "TESTINIT14"
}